Update tree-sitter to latest version
I was following this [guide]() to create a Java parser and I got this error:
"called `Result::unwrap()` on an `Err` value: "Incompatible language version 13. Expected 10.""
following https://github.com/tree-sitter/tree-sitter/issues/966 I noticed that the problem was the tree-sitter version.
I updated the version locally to the latest plugin version (0.20.9) and now it works.